ms-property/project/pyproject.toml
[tool.poetry]
name = "ms-property"
version = "0.1.0"
description = ""
authors = ["Diego Quiroz <diego00aq@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
mysql-connector-python = "^8.0.29"
pytest-httpserver = "^1.0.4"
gunicorn = "^20.1.0"
WebOb = "^1.8.7"
requests-wsgi-adapter = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | wsgi.py
  | manage.py
)/
'''
